Splitted `PackageWriter` into multiple generators
