+++ build.gradle
@@ -21,6 +21,7 @@ repositories {
 dependencies {
     compile "org.jetbrains.kotlin:kotlin-stdlib-jre8:$kotlin_version"
     compile 'com.google.code.gson:gson:2.8.2'
+    compile 'com.squareup:javapoet:1.9.0'
 
     testCompile 'junit:junit:4.12'
     testCompile "org.mockito:mockito-inline:2.10.0"
