+++ src/main/kotlin/ui/generators/BuildGradleGenerator.kt
@@ -1,11 +1,12 @@
-package ui
+package ui.generators
 
+import ui.BuildGradle
 import ui.models.ModuleBlueprint
 
-class BuildGradleCreator {
+class BuildGradleGenerator {
     fun create(moduleBlueprint: ModuleBlueprint): String {
         return BuildGradle.print(moduleBlueprint.dependencies
                 .map { it -> "compile project(':module$it')\n" }
-                .fold("", {acc, next -> acc + next }))
+                .fold("", { acc, next -> acc + next }))
     }
 }
